+++ integration_tests/suite/base/test_line_endpoint_custom.py
@@ -1,7 +1,12 @@
 # Copyright 2016-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from hamcrest import assert_that, has_entries
+from hamcrest import (
+    assert_that,
+    greater_than,
+    has_entries,
+    none,
+)
 
 from . import confd
 from ..helpers import (
@@ -10,7 +15,7 @@ from ..helpers import (
     fixtures,
     scenarios as s,
 )
-from ..helpers.config import MAIN_TENANT, SUB_TENANT
+from ..helpers.config import CONTEXT, MAIN_TENANT, SUB_TENANT
 
 
 @fixtures.line()
@@ -203,3 +208,55 @@ def test_bus_events(line, custom):
 
     yield s.check_event, 'line_endpoint_custom_associated', headers, url.put
     yield s.check_event, 'line_endpoint_custom_dissociated', headers, url.delete
+
+
+@fixtures.registrar()
+def test_create_line_with_endpoint_custom_with_all_parameters(registrar):
+    response = confd.lines.post(
+        context=CONTEXT,
+        position=2,
+        registrar=registrar['id'],
+        provisioning_code='887865',
+        endpoint_custom={
+            'interface': 'custom/createall',
+            'enabled': False,
+        },
+    )
+
+    try:
+        assert_that(
+            response.item,
+            has_entries(
+                context=CONTEXT,
+                position=2,
+                device_slot=2,
+                name='custom/createall',
+                protocol='custom',
+                device_id=none(),
+                caller_id_name=none(),
+                caller_id_num=none(),
+                registrar=registrar['id'],
+                provisioning_code="887865",
+                provisioning_extension="887865",
+                tenant_uuid=MAIN_TENANT,
+                endpoint_custom=has_entries(
+                    id=greater_than(0),
+                    tenant_uuid=MAIN_TENANT,
+                    interface='custom/createall',
+                    enabled=False,
+                ),
+            ),
+        )
+
+        assert_that(
+            confd.lines(response.item['id']).get().item,
+            has_entries(
+                endpoint_custom=has_entries(
+                    id=response.item['endpoint_custom']['id'],
+                    interface='custom/createall',
+                ),
+            ),
+        )
+
+    finally:
+        confd.lines(response.item['id']).delete().assert_deleted()