+++ integration_tests/suite/base/test_line_endpoint_sccp.py
@@ -1,7 +1,14 @@
 # Copyright 2015-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from hamcrest import assert_that, has_entries
+from hamcrest import (
+    assert_that,
+    contains_inanyorder,
+    has_entries,
+    greater_than,
+    none,
+    not_,
+)
 
 from . import confd
 from ..helpers import (
@@ -10,7 +17,7 @@ from ..helpers import (
     fixtures,
     scenarios as s,
 )
-from ..helpers.config import MAIN_TENANT, SUB_TENANT
+from ..helpers.config import CONTEXT, MAIN_TENANT, SUB_TENANT
 
 
 @fixtures.line()
@@ -187,3 +194,61 @@ def test_bus_events(line, sccp):
 
     yield s.check_event, 'line_endpoint_sccp_associated', headers, url.put
     yield s.check_event, 'line_endpoint_sccp_dissociated', headers, url.delete
+
+
+@fixtures.registrar()
+def test_create_line_with_endpoint_sccp_with_all_parameters(registrar):
+    response = confd.lines.post(
+        context=CONTEXT,
+        position=2,
+        registrar=registrar['id'],
+        provisioning_code='887865',
+        endpoint_sccp={
+            'options': [
+                ["cid_name", "cid_name"],
+                ["cid_num", "cid_num"],
+                ["allow", "allow"],
+                ["disallow", "disallow"],
+            ],
+        },
+    )
+
+    try:
+        line_id = response.item['id']
+        endpoint_sccp_id = response.item['endpoint_sccp']['id']
+        assert_that(
+            response.item,
+            has_entries(
+                context=CONTEXT,
+                position=2,
+                device_slot=2,
+                name=not_(none()),
+                protocol='sccp',
+                device_id=none(),
+                caller_id_name='cid_name',
+                caller_id_num='cid_num',
+                registrar=registrar['id'],
+                provisioning_code="887865",
+                provisioning_extension="887865",
+                tenant_uuid=MAIN_TENANT,
+                endpoint_sccp=has_entries(
+                    id=greater_than(0),
+                    options=contains_inanyorder(
+                        ["cid_name", "cid_name"],
+                        ["cid_num", "cid_num"],
+                        ["allow", "allow"],
+                        ["disallow", "disallow"],
+                    ),
+                ),
+            ),
+        )
+
+        assert_that(
+            confd.lines(line_id).get().item,
+            has_entries(
+                endpoint_sccp=has_entries(id=endpoint_sccp_id),
+            ),
+        )
+
+    finally:
+        confd.lines(response.item['id']).delete().assert_deleted()