+++ wazo_confd/plugins/endpoint_custom/plugin.py
@@ -1,6 +1,7 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from .middleware import EndpointCustomMiddleWare
 from .resource import CustomItem, CustomList
 from .service import build_service
 
@@ -8,7 +9,13 @@ from .service import build_service
 class Plugin:
     def load(self, dependencies):
         api = dependencies['api']
+        middleware_handle = dependencies['middleware_handle']
+
         service = build_service()
+
+        endpoint_custom_middleware = EndpointCustomMiddleWare(service)
+
+        middleware_handle.register('endpoint_custom', endpoint_custom_middleware)
 
         api.add_resource(
             CustomItem,
@@ -17,5 +24,7 @@ class Plugin:
             resource_class_args=(service,),
         )
         api.add_resource(
-            CustomList, '/endpoints/custom', resource_class_args=(service,)
+            CustomList,
+            '/endpoints/custom',
+            resource_class_args=(service, endpoint_custom_middleware),
         )