+++ wazo_confd/plugins/endpoint_sip/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2021 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_provd_client import Client as ProvdClient
@@ -8,6 +8,7 @@ from xivo_dao.resources.pjsip_transport import dao as transport_dao
 
 from .resource import SipItem, SipList, SipTemplateItem, SipTemplateList
 from .service import build_endpoint_service, build_template_service
+from .middleware import EndpointSIPMiddleWare, TemplateSIPMiddleWare
 
 
 class Plugin:
@@ -15,6 +16,7 @@ class Plugin:
         api = dependencies['api']
         config = dependencies['config']
         token_changed_subscribe = dependencies['token_changed_subscribe']
+        middleware_handle = dependencies['middleware_handle']
 
         provd_client = ProvdClient(**config['provd'])
         token_changed_subscribe(provd_client.set_token)
@@ -22,6 +24,11 @@ class Plugin:
 
         endpoint_service = build_endpoint_service(provd_client, pjsip_doc)
         template_service = build_template_service(provd_client, pjsip_doc)
+
+        endpoint_sip_middleware = EndpointSIPMiddleWare(endpoint_service)
+        template_sip_middleware = TemplateSIPMiddleWare(template_service)
+
+        middleware_handle.register('endpoint_sip', endpoint_sip_middleware)
 
         api.add_resource(
             SipItem,
@@ -32,7 +39,12 @@ class Plugin:
         api.add_resource(
             SipList,
             '/endpoints/sip',
-            resource_class_args=(endpoint_service, sip_dao, transport_dao),
+            resource_class_args=(
+                endpoint_service,
+                sip_dao,
+                transport_dao,
+                endpoint_sip_middleware,
+            ),
         )
         api.add_resource(
             SipTemplateItem,
@@ -43,5 +55,10 @@ class Plugin:
         api.add_resource(
             SipTemplateList,
             '/endpoints/sip/templates',
-            resource_class_args=(template_service, sip_dao, transport_dao),
+            resource_class_args=(
+                template_service,
+                sip_dao,
+                transport_dao,
+                template_sip_middleware,
+            ),
         )