+++ wazo_confd/plugins/line/plugin.py
@@ -1,8 +1,9 @@
-# Copyright 2015-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_provd_client import Client as ProvdClient
 
+from .middleware import LineMiddleWare
 from .resource import LineItem, LineList
 from .service import build_service
 
@@ -12,11 +13,14 @@ class Plugin:
         api = dependencies['api']
         config = dependencies['config']
         token_changed_subscribe = dependencies['token_changed_subscribe']
+        middleware_handle = dependencies['middleware_handle']
 
         provd_client = ProvdClient(**config['provd'])
         token_changed_subscribe(provd_client.set_token)
 
         service = build_service(provd_client)
+
+        line_middleware = LineMiddleWare(service, middleware_handle)
 
         api.add_resource(
             LineItem,
@@ -24,4 +28,8 @@ class Plugin:
             endpoint='lines',
             resource_class_args=(service,),
         )
-        api.add_resource(LineList, '/lines', resource_class_args=(service,))
+        api.add_resource(
+            LineList,
+            '/lines',
+            resource_class_args=(service, line_middleware),
+        )