+++ wazo_confd/plugins/line_endpoint/plugin.py
@@ -1,10 +1,5 @@
-# Copyright 2015-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
-
-from xivo_dao.resources.endpoint_custom import dao as endpoint_custom_dao
-from xivo_dao.resources.endpoint_sccp import dao as endpoint_sccp_dao
-from xivo_dao.resources.endpoint_sip import dao as endpoint_sip_dao
-from xivo_dao.resources.line import dao as line_dao
 
 from wazo_provd_client import Client as ProvdClient
 
@@ -18,6 +13,11 @@ from .service import (
     build_service_sccp,
     build_service_custom,
 )
+from .middleware import (
+    LineEndpointCustomMiddleWare,
+    LineEndpointSCCPMiddleWare,
+    LineEndpointSIPMiddleWare,
+)
 
 
 class Plugin:
@@ -25,6 +25,7 @@ class Plugin:
         api = dependencies['api']
         config = dependencies['config']
         token_changed_subscribe = dependencies['token_changed_subscribe']
+        middleware_handle = dependencies['middleware_handle']
 
         provd_client = ProvdClient(**config['provd'])
         token_changed_subscribe(provd_client.set_token)
@@ -33,21 +34,30 @@ class Plugin:
         service_sccp = build_service_sccp(provd_client)
         service_custom = build_service_custom(provd_client)
 
+        line_endpoint_custom_middleware = LineEndpointCustomMiddleWare(service_custom)
+        line_endpoint_sccp_middleware = LineEndpointSCCPMiddleWare(service_sccp)
+        line_endpoint_sip_middleware = LineEndpointSIPMiddleWare(service_sip)
+        middleware_handle.register(
+            'line_endpoint_custom', line_endpoint_custom_middleware
+        )
+        middleware_handle.register('line_endpoint_sccp', line_endpoint_sccp_middleware)
+        middleware_handle.register('line_endpoint_sip', line_endpoint_sip_middleware)
+
         api.add_resource(
             LineEndpointAssociationSip,
             '/lines/<int:line_id>/endpoints/sip/<uuid:endpoint_uuid>',
             endpoint='line_endpoint_sip',
-            resource_class_args=(service_sip, line_dao, endpoint_sip_dao),
+            resource_class_args=(line_endpoint_sip_middleware,),
         )
         api.add_resource(
             LineEndpointAssociationSccp,
             '/lines/<int:line_id>/endpoints/sccp/<int:endpoint_id>',
             endpoint='line_endpoint_sccp',
-            resource_class_args=(service_sccp, line_dao, endpoint_sccp_dao),
+            resource_class_args=(line_endpoint_sccp_middleware,),
         )
         api.add_resource(
             LineEndpointAssociationCustom,
             '/lines/<int:line_id>/endpoints/custom/<int:endpoint_id>',
             endpoint='line_endpoint_custom',
-            resource_class_args=(service_custom, line_dao, endpoint_custom_dao),
+            resource_class_args=(line_endpoint_custom_middleware,),
         )