+++ wazo_confd/plugins/line_endpoint/resource.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_confd.auth import required_acl
@@ -9,47 +9,29 @@ class LineEndpointAssociation(ConfdResource):
 
     has_tenant_uuid = True
 
-    def __init__(self, service, line_dao, endpoint_dao):
+    def __init__(self, middleware):
         super().__init__()
-        self.service = service
-        self.line_dao = line_dao
-        self.endpoint_dao = endpoint_dao
+        self._middleware = middleware
 
     def put(self, line_id, endpoint_id):
         tenant_uuids = self._build_tenant_list({'recurse': True})
-        line = self.line_dao.get(line_id, tenant_uuids=tenant_uuids)
-        endpoint = self.endpoint_dao.get(endpoint_id, tenant_uuids=tenant_uuids)
-        self.service.associate(line, endpoint)
+        self._middleware.associate(line_id, endpoint_id, tenant_uuids)
         return '', 204
 
     def delete(self, line_id, endpoint_id):
         tenant_uuids = self._build_tenant_list({'recurse': True})
-        line = self.line_dao.get(line_id, tenant_uuids=tenant_uuids)
-        endpoint = self.endpoint_dao.get(endpoint_id, tenant_uuids=tenant_uuids)
-        self.service.dissociate(line, endpoint)
+        self._middleware.dissociate(line_id, endpoint_id, tenant_uuids)
         return '', 204
 
 
 class LineEndpointAssociationSip(LineEndpointAssociation):
     @required_acl('confd.lines.{line_id}.endpoints.sip.{endpoint_uuid}.update')
     def put(self, line_id, endpoint_uuid):
-        tenant_uuids = self._build_tenant_list({'recurse': True})
-        line = self.line_dao.get(line_id, tenant_uuids=tenant_uuids)
-        sip = self.endpoint_dao.get(
-            endpoint_uuid, template=False, tenant_uuids=tenant_uuids
-        )
-        self.service.associate(line, sip)
-        return '', 204
+        return super().put(line_id, endpoint_uuid)
 
     @required_acl('confd.lines.{line_id}.endpoints.sip.{endpoint_uuid}.delete')
     def delete(self, line_id, endpoint_uuid):
-        tenant_uuids = self._build_tenant_list({'recurse': True})
-        line = self.line_dao.get(line_id, tenant_uuids=tenant_uuids)
-        sip = self.endpoint_dao.get(
-            endpoint_uuid, template=False, tenant_uuids=tenant_uuids
-        )
-        self.service.dissociate(line, sip)
-        return '', 204
+        return super().delete(line_id, endpoint_uuid)
 
 
 class LineEndpointAssociationSccp(LineEndpointAssociation):