+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 22.15
 
 * The `/tenants` endpoints were modified to remove the field `webrtc_video_sip_template_uuid`.
+* `/lines` endpoint now create the SIP, SCCP or custom endpoint if it is specified in the body and associates it to the line.
 
 ## 22.14